+++ src/modes/loopTypewriter.js
@@ -1,6 +1,7 @@
 import { writeEffect } from '../helpers/writeEffect'
 import { unwriteEffect } from '../helpers/unwriteEffect'
 import { getRandomElement } from '../helpers/getRandomElement'
+import { runOnEveryParentUntil } from '../helpers/runOnEveryParentUntil'
 
 /** @type {import('types').TypewriterEffectFn} */
 const loopTypewriter = async ({ currentNode, text }, options) => {
@@ -8,6 +9,11 @@ const loopTypewriter = async ({ currentNode, text }, options) => {
 	const textWithUnescapedAmpersands = text.replaceAll('&', '&amp;')
 	const fullyWritten = currentNode.innerHTML === textWithUnescapedAmpersands
 	fullyWritten && (await unwriteEffect(currentNode, options))
+	runOnEveryParentUntil(currentNode, options.parentElement, element => {
+		currentNode === element
+			? element.classList.remove('typing')
+			: element.classList.remove('finished-typing')
+	})
 }
 
 /** @type {import('types').TypewriterModeFn} */