+++ src/modes/typewriter.js
@@ -11,8 +11,17 @@ const mode = async (elements, options) => {
 		const lastElementToFinish = getLongestTextElement(elements)
 		onAnimationEnd(lastElementToFinish, () => options.dispatch('done'))
 	}
-	for (const element of elements)
-		options.cascade ? await writeEffect(element, options) : writeEffect(element, options)
+	for (const element of elements) {
+		if (options.cascade) {
+			await writeEffect(element, options)
+			element.currentNode.classList.replace('typing', 'finished-typing')
+		} else {
+			writeEffect(element, options).then(() => {
+				element.currentNode.classList.replace('typing', 'finished-typing')
+			})
+		}
+	}
+
 	options.cascade && options.dispatch('done')
 }